segmenter/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-multipart==0.0.20
numpy==2.2.5
Pillow==11.2.1
